requirements.txt
#basic env
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
colorful-logger
influxdb-client
paho-mqtt
apscheduler
#dev env 
pylint
pyyaml
rich
autopep8